加入注释的版本，来自https://github.com/lwjaiyjk/bitcoin-comment-0.1.0

+++ bitcoin-comment-0.1/src/db.cpp
@@ -15,7 +15,8 @@
 
 static CCriticalSection cs_db;
 static bool fDbEnvInit = false;
-DbEnv dbenv(0);
+//DbEnv dbenv(0);
+DbEnv dbenv((u_int32_t)0);
 static map<string, int> mapFileUseCount;
 
 class CDBInit
